Always use system python.
People using a python installed by homebrew or macports were having problems
as PyObjC wasn't in the python_path for their python. None of the package
managers mess with the system python installed /usr/bin so we'll use that
instead of whatever happens to be in $PATH which should make everything just
work for everyone.