Support for webkit-transform (thanks Nate Weaver)
As noted in https://github.com/paulhammond/webkit2png/issues/13, the method
we used to capture the contents of the window didn't work with
webkit-transform. Nate Weaver (@wevar) provided some nifty code that should
fix this when run on Mac OS X 10.5 or later whilst still letting the script
run on older versions.

Only problem is that I've only been able to test this on 10.7, 10.6 and 10.5.
I don't think many people are running webkit2png on older OS versions so this
should be OK.