Divide --cookies into --cookie-file and --cookie
--cookie-file is the old behavior, where you pass in a filename (or
a FIFO). --cookie is used to pass in key-value pairs, either in a
single string, or with multiple --cookie options:

--cookie 'name1=value1; name2=value2; name3=value3'

...is the same as:

--cookie name1=value1 --cookie name2=value2 --cookie name3=value3

Both --cookie and --cookie-file are allowed at the same time, in
which case --cookie values are blindly appended to the end of the
cookies that are parsed out of the cookie file.